pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "load-assurance"
version = "0.1.0"
description = "Pluggable, metadata-first data load validation across platforms"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",        # pretty CLI output
    "pydantic>=2.0",     # config validation
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock", "moto[s3]>=5.0", "ruff"]

[project.scripts]
la = "load_assurance.cli:main"

[tool.setuptools.packages.find]
include = ["load_assurance*"]